+++ lib/passive_liveness_detection.dart
@@ -2,5 +2,7 @@ library passive_liveness_detection;
 
 export 'src/liveness_detector.dart';
 export 'src/liveness_result.dart';
+export 'src/liveness_window.dart';
 export 'src/face_box.dart';
+export 'src/face_framing_gate.dart';
 export 'src/crop_utils.dart';